+++ lib/karafka/attributes_map.rb
@@ -49,6 +49,7 @@ module Karafka
           parser
           interchanger
           responder
+          batch_processing
         ]).uniq
       end
 
@@ -62,7 +63,7 @@ module Karafka
         #   only when proxying details go ruby-kafka. We use ignored fields internally in karafka
         ignored_settings = config_adapter[:subscription]
         defined_settings = config_adapter.values.flatten
-        karafka_settings = %i[batch_consuming batch_processing topic_mapper]
+        karafka_settings = %i[batch_consuming topic_mapper]
         # This is a drity and bad hack of dry-configurable to get keys before setting values
         dynamically_proxied = Karafka::Setup::Config
                               ._settings