+++ lib/karafka/errors.rb
@@ -4,12 +4,12 @@ module Karafka
   # Namespace used to encapsulate all the internal errors of Karafka
   module Errors
     # Base class for all the Karafka internal errors
-    class BaseError < StandardError; end
+    BaseError = Class.new(StandardError)
 
     # Should be raised when we attemp to parse incoming params but parsing fails
     #   If this error (or its descendant) is detected, we will pass the raw message
     #   into params and proceed further
-    class ParserError < BaseError; end
+    ParserError = Class.new(BaseError)
 
     # Raised when router receives topic name which does not correspond with any routes
     # This can only happen in a case when:
@@ -21,17 +21,21 @@ module Karafka
     # In case this happens, you will have to create a temporary route that will allow
     # you to "eat" everything from the Sidekiq queue.
     # @see https://github.com/karafka/karafka/issues/135
-    class NonMatchingRouteError < BaseError; end
+    NonMatchingRouteError = Class.new(BaseError)
 
     # Raised when application does not have ApplicationWorker or other class that directly
     # inherits from Karafka::BaseWorker
-    class BaseWorkerDescentantMissing < BaseError; end
+    BaseWorkerDescentantMissing = Class.new(BaseError)
 
     # Raised when we want to use #respond_with in controllers but we didn't define
     # (and we couldn't find) any appropriate responder for a given controller
-    class ResponderMissing < BaseError; end
+    ResponderMissing = Class.new(BaseError)
 
     # Raised when configuration doesn't match with validation schema
-    class InvalidConfiguration < BaseError; end
+    InvalidConfiguration = Class.new(BaseError)
+
+    # Raised when processing messages in batches but still want to use #params instead of
+    # #params_batch
+    ParamsMethodUnavailable = Class.new(BaseError)
   end
 end