+++ spec/lib/karafka/attributes_map_spec.rb
@@ -22,6 +22,7 @@ RSpec.describe Karafka::AttributesMap do
         parser
         interchanger
         responder
+        batch_processing
       ]
     end
 
@@ -40,7 +41,6 @@ RSpec.describe Karafka::AttributesMap do
 
     it 'expect to include karafka specific settings' do
       expect(map.consumer_group).to include(:batch_consuming)
-      expect(map.consumer_group).to include(:batch_processing)
       expect(map.consumer_group).to include(:topic_mapper)
     end
   end