+++ spec/lib/karafka/fetcher_spec.rb
@@ -67,7 +67,7 @@ RSpec.describe Karafka::Fetcher do
 
     context 'when we invoke a processor block' do
       let(:message) { double }
-      let(:processor) { Karafka::Connection::MessageProcessor }
+      let(:processor) { Karafka::Connection::MessagesProcessor }
       let(:consumer_group_id) { rand.to_s }
 
       it 'process the message' do